+++ Examples/Resources/SqlServerRole/1-AddServerRole.ps1
@@ -1,7 +1,8 @@
 <#
 .EXAMPLE
-    This example shows how to ensure that the server role named
-    AdminSqlforBI is present on instance sqltest.company.local\DSC.
+    This example shows how to ensure that both the server role named
+    MyServerRole1 and MyServerRole2 is present on instance
+    'sqltest.company.local\DSC'.
 #>
 
 Configuration Example
@@ -9,19 +10,28 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
-        $SysAdminAccount
+        $SqlAdministratorCredential
     )
 
     Import-DscResource -ModuleName SqlServerDsc
 
     node localhost {
-        SqlServerRole Add_ServerRole_AdminSqlforBI
+        SqlServerRole Add_ServerRole_MyServerRole1
         {
             Ensure               = 'Present'
-            ServerRoleName       = 'AdminSqlforBI'
+            ServerRoleName       = 'MyServerRole1'
             ServerName           = 'sqltest.company.local'
             InstanceName         = 'DSC'
-            PsDscRunAsCredential = $SysAdminAccount
+            PsDscRunAsCredential = $SqlAdministratorCredential
+        }
+
+        SqlServerRole Add_ServerRole_MyServerRole2
+        {
+            Ensure               = 'Present'
+            ServerRoleName       = 'MyServerRole2'
+            ServerName           = 'sqltest.company.local'
+            InstanceName         = 'DSC'
+            PsDscRunAsCredential = $SqlAdministratorCredential
         }
     }
 }